Overrode the get_name() function (#7)
* Overrode the get_name() function

Perhaps the Human Classes version of get_name() was returning this->name which somehow was being confused for the pImpl_ pointer?
I noticed that pImpl_->get_name() was actually returning the correct name and this->get_name() and this->get_name() for both the Human object and
BaseballplayerImpl object was returning the correct name value but this->get_name() for the BaseballPlayer object was not. Perhaps this is not the most ideal
solution but then again pimpl isnt always the most ideal solution either just trying to use this time to learn. I will continue to read in Herb Sutters book
"Exceptional C++: 47 engineering puzzles" He has chapters on pimpl and forward declarations

Let me know if you find a more elegant solution

* removing iostream headers